Embedded C++ code into webpage via Web Assembly
This branch is in the form of a webpage (index.html and index.js) that imports a Web Assembly module that has been created from the main.cpp source code.  This cpp source code just defines a single function called generateMessage. The Web Assembly toolchain has been used to compile this down to output.wasm and a wrapper output.js that can be loaded into the webpage.